Created a SpaceGameArea class by duplicating the ForestGameArea class which extends the GameArea class. This class will be used similarly to the ForestGameArea class and used to control the map and terrain. Currently the only difference is that it spawns more enemies.
In the TerrainFactory class, a new function was created to create a TiledMap instance createSpaceGameTiles. This function will be used to generate a TiledMap instance (or a new MapClass which extends TiledMap) which will be used to draw the terrain. It is still being developed and is not called in the code yet.

In the MainGameScreen class, references to the ForestGameArea class are replaced with the SpaceGameArea class. This replaces the Forest Game demo with the Space Game we are making (which is now identical to the demo except for more monsters).