Created new Class GameMap to store TerrainFactory and TiledMap of the game. Added getter methods to access these values and a function to getTiles from the TiledMap. Also added an EventHandler variable in case it is needed for future use.
Updated the TerrainFactory and SpaceGameArea classes to now include an instance of the GameMap class. These classes also have been modified to make use of this GameMap instance, so that values within the GameMap class can be used by functons in the SpaceGameArea. The ForestGameArea has also been modified slightly so that I did not have to comment out the entire class due to functions being modified for use in the SpaceGameArea class.